requirements.txt
pydantic-settings
fastapi~=0.116.1
pillow~=11.3.0
uvicorn~=0.35.0
pydantic~=2.11.7
python-multipart
